REFACTOR : problemId의 타입을 long에서 int로 변경

+++ src/main/java/com/algoanalyzer/domain/analysis/problem/dto/request/ProblemAnalysisRequestDto.java
@@ -11,11 +11,11 @@ import java.util.List;
 public class ProblemAnalysisRequestDto {
     @NotNull(message = "문제 번호는 필수입니다.")
     @JsonProperty("problem_id")
-    private Long problemId;
+    private int problemId;
     
     @JsonProperty("description")
     private String description;
-    
+
     @JsonProperty("input")
     private String input;
